+++ backend/src/main/java/com/ssafy/patpat/shelter/dto/BreedDto.java
@@ -11,13 +11,13 @@ import lombok.*;
 @Builder
 public class BreedDto {
     @Schema(example = "견종id")
-    private int breedId;
+    private Long breedId;
     @Schema(example = "견종")
     private String breedName;
     @Schema(example = "fileUrl:dd.png")
     private FileDto fileDto;
     @Schema(example = "분양횟수")
-    private int adoptCount;
+    private Integer adoptCount;
     @Schema(example = "제목")
     private String title;
     @Schema(example = "내용")
